Revert "Move serves_https_requests_with_cert_from_cache_directory"
This reverts commit 47b3287f356c8bd2e75526db16db19fac1f10bc5.
